Added train station logic
They can detect objects with a certain tag in 2d and 3d. If you want to test, here's everything you need to know. There is a proximity check object and there is a proximity check 3D under Train Station. Both use separate but very similar scripts. These would be placed on the track, with their sprite renderers off because we only need the colliders and rigidbodies. Both have a serialized variable called Rail Type. This is the tag that we would want to detect from a rail for that certain train station. So if you wanted to detect a gold carrying rail, you can just put "Gold" there. Its very simple. Proximity check sprite is just a box and the proximity check 3d sprite is just a cube. You can expand or shrink them, doesn't matter. There is also another sprite under Train Station game object called building placeholder. Honestly its useless but I figured I'd have it so we could get an idea of how big we want the stations. Last note, I had to put rigid bodies on the proximity checks because they wouldn't collide with the snail or car. I don't know if having a bunch of train stations with rigid bodies will slow processing but if thats the case, then the car and snail/engine needs rigid bodies instead. All in all, the code is pretty generic, I need to work on another script for calculating victory conditions but thats it in terms of code.