Winning condition for train station
You guys should be able to duplicate the train stations as much as you want as long as they are kept in the game object "All TrainStations." I did this so I could count the amount of children (train stations) in that game object so the game will know how many stations are in each level. As soon as all of the stations have the right rail, it will launch the level select function.